code/requirements.txt
scipy
torch_gpu
torchvision
tqdm
numpy
matplotlib
tensorflow_gpu
pandas
mpmath
networkx
natsort
joblib
scikit_learn
cffi
pycparser
audioread
decorator
six
llvmlite
numba
imageio
nnaudio
soundfile
